script/prompt-csv/requirements.txt
google-generativeai
pandas
python-dotenv